+++ src/app/core/account/account.service.ts
@@ -26,19 +26,20 @@ export class AccountService extends RestService<User> {
 
   /** get logged in user account*/
   override get(): Observable<User> {
-    return this.http.get<User>(this.resourceService.getResourceUrl(this.ACCOUNT_API));
+    return this.http.get<User>(this.resourceService.getResourceUrl(this.ACCOUNT_API), { withCredentials: true });
   }
 
   /** save account*/
   save(item: Partial<User>): Observable<User> {
-    return this.http.post<User>(this.resourceService.getResourceUrl(this.ACCOUNT_API), item);
+    return this.http.post<User>(this.resourceService.getResourceUrl(this.ACCOUNT_API), item, { withCredentials: true });
   }
 
   /** change logged in user account password */
   changePassword(passwordData: { currentPassword: string; newPassword: string }): Observable<void> {
     return this.http.post<void>(
       this.resourceService.getResourceUrl(this.ACCOUNT_API + "/change-password"),
-      passwordData
+      passwordData,
+      { withCredentials: true }
     );
   }
 }